Fix ModuleNotFoundError for auth import
