+++ meetup_clone/app/routes.py
@@ -1,13 +1,12 @@
 from flask import Blueprint, render_template, redirect, url_for, flash, request
-from flask_login import login_required, current_user, login_user, logout_user
-from app.models import Event, User
-from app.forms import EventForm, RegisterForm, LoginForm
+from flask_login import login_required, current_user
+from app.models import Event
+from app.forms import EventForm
 from app import db
 from datetime import datetime
 import random
 
 main = Blueprint('main', __name__)
-auth = Blueprint('auth', __name__)
 
 @main.route('/')
 def home():
@@ -49,35 +48,3 @@ def create_event():
 def event_details(event_id):
     event = Event.query.get_or_404(event_id)
     return render_template('event_details.html', event=event)
-
-@auth.route('/register', methods=['GET', 'POST'])
-def register():
-    form = RegisterForm()
-    if form.validate_on_submit():
-        user = User(username=form.username.data, email=form.email.data, phone=form.phone.data)
-        user.set_password(form.password.data)
-        db.session.add(user)
-        db.session.commit()
-        flash('Registration successful! You can now log in.', 'success')
-        return redirect(url_for('auth.login'))
-    return render_template('register.html', form=form)
-
-@auth.route('/login', methods=['GET', 'POST'])
-def login():
-    form = LoginForm()
-    if form.validate_on_submit():
-        user = User.query.filter_by(email=form.email.data).first()
-        if user and user.check_password(form.password.data):
-            login_user(user)
-            flash('Login successful!', 'success')
-            return redirect(url_for('main.home'))
-        else:
-            flash('Invalid email or password.', 'danger')
-    return render_template('login.html', form=form)
-
-@auth.route('/logout')
-@login_required
-def logout():
-    logout_user()
-    flash('You have been logged out.', 'success')
-    return redirect(url_for('main.home'))
